Forms can be created and sent per email. TAN invalidation also successful after pdf creation.

+++ functions/mercApi/apiMain.py
@@ -8,20 +8,23 @@ import typing
 from threading import Lock
 import secrets
 
-tanListLock = Lock()
+tanListFileLock = Lock()
+tanListAppLock = Lock()
 tanListFileName = "tanList.json"
 tanDictAppConfigKey = "TAN_DICT"
 tanDictEmailKey = "tanDictEmailKey"
 tanDictIndexKey = "tanDictIndexKey"
-TAN_LIST_SIZE = 128
-TAN_LIST_RENEWAL_RETAIN_SIZE = 16
+tanDictLargestIndexKey = "tanDictLargestIndex"
+tanDictFormEnumKey = "tanDictFormEnumKey"
+TAN_LIST_SIZE = 4
+TAN_LIST_RENEWAL_RETAIN_SIZE = 2
 app = Flask(__name__)
 CORS(app)
 
 def send_email_with_attachment(subject: str,
                                content: str,
                                from_field: str = "",
-                               fileNameFilePathDict: typing.Optional[typing.Dict[str, str]] = None,
+                               fileNameFilePathDict: typing.Optional[typing.Dict[str, typing.Union[str, bytes]]] = None,
                                deleteFiles: bool = True) -> bool:
     if app.config["YOUR_STRATO_USER"] is None or app.config["YOUR_STRATO_PASS"] is None:
         return False
@@ -30,16 +33,18 @@ def send_email_with_attachment(subject: str,
     msg['Subject'] = subject
     msg['From'] = "stv.schnelllauf@merc-online.de"
     msg['To'] = "rastapopoulis@hotmail.com" # change this for deployment
-    if isinstance(from_field, str) and "@" in from_field and from_field.count("@") == 1:
+    if isinstance(from_field, str) and "@" in from_field and from_field.count("@") == 1 and from_field != msg['From'] and from_field != msg['To']:
         msg['cc'] = from_field
     msg.set_content(content)
 
     try:
         if fileNameFilePathDict is not None:
             for filename, file_path in fileNameFilePathDict.items():
-                with open(file_path, 'rb') as f:
-                    file_data = f.read()
-                    msg.add_attachment(
+                file_data = file_path
+                if isinstance(file_path, str):
+                    with open(file_path, 'rb') as f:
+                        file_data = f.read()
+                msg.add_attachment(
                         file_data,
                         maintype='application',
                         subtype='octet-stream',
@@ -56,9 +61,9 @@ def send_email_with_attachment(subject: str,
             server.send_message(msg)
             
         print("Email sent successfully")
-        if deleteFiles:
+        if deleteFiles and fileNameFilePathDict is not None:
             for filename, file_path in fileNameFilePathDict.items():
-                if os.path.exists(file_path) and os.path.isfile(file_path):
+                if isinstance(file_path, str) and os.path.exists(file_path) and os.path.isfile(file_path):
                     os.remove(file_path)
         return True
     except Exception as error:
@@ -69,11 +74,12 @@ def _generateTanListHelper() -> typing.Dict[str, typing.Dict[str, typing.Union[i
     TAN_DICT = {}
     for _ind in range(TAN_LIST_SIZE):
         hash_hex = secrets.token_hex(nbytes=3)
-        TAN_DICT[hash_hex] = {tanDictIndexKey:0, tanDictEmailKey:""}
+        TAN_DICT[hash_hex] = {tanDictIndexKey:0, tanDictEmailKey:"", tanDictFormEnumKey: '-1'}
+    TAN_DICT[tanDictLargestIndexKey] = 0
     return TAN_DICT
 
 def _generateTanList(TAN_DICT: typing.Optional[typing.Dict[str, typing.Dict[str, typing.Union[int,str]]]] = None) -> typing.Tuple[typing.Dict[str, typing.Union[str, bool]], int]:
-    with tanListLock:
+    with tanListFileLock:
         try:
             if os.path.exists(os.path.join(os.path.dirname(__file__), tanListFileName)):
                 os.remove(os.path.join(os.path.dirname(__file__), tanListFileName))
@@ -109,9 +115,9 @@ if not os.path.exists(os.path.join(os.path.dirname(__file__), tanListFileName)):
         print(out)
 
 app.config[tanDictAppConfigKey] = {}
-
-with open(os.path.join(os.path.dirname(__file__), tanListFileName), "r") as f:
-    app.config[tanDictAppConfigKey] = json.load(f)
+with tanListFileLock:
+    with open(os.path.join(os.path.dirname(__file__), tanListFileName), "r") as f:
+        app.config[tanDictAppConfigKey] = json.load(f)
 
 @app.route('/generateNewTanList', methods=["GET"])
 def generateTanList() -> typing.Tuple[typing.Dict[str, typing.Union[str, bool]],int]:
@@ -125,13 +131,21 @@ def submit_form() -> typing.Tuple[typing.Dict[str, typing.Union[str, bool]],int]
         if request.method != "POST":
             return jsonify({"success":False, "error": "request.method nicht vom Typ POST"}), 400
         file = request.form.get('attachment', None)
+        fileAsBlob = request.files.get('attachment', None)
         subject = request.form.get("subject", "Unbekannt")
         content = request.form.get("message", "Leer")
         from_field = request.form.get("email", "stv.schnelllauf@merc-online.de")
-        filename = None
+        attachmentDict = None
         if file is not None:
-            filename = os.path.basename(file)
-        out = send_email_with_attachment(subject, content, from_field, {filename: file})
+            if isinstance(file, str):
+                filename = os.path.basename(file)
+                attachmentDict = {filename: file}
+        if fileAsBlob is not None:
+            explicitFileName = request.form.get('attachmentName', None)
+            if explicitFileName is None:
+                return jsonify({"success": False, "error":"Param 'attachmentName' fehlt im Request"}), 400
+            attachmentDict = {explicitFileName: fileAsBlob.read()}
+        out = send_email_with_attachment(subject, content, from_field, attachmentDict)
         if out:
             return jsonify({"success": True, "message": "Nachricht gesendet"}), 200
     except Exception as e:
@@ -146,7 +160,8 @@ def isTanValid() -> str:
        Return '2': TAN gueltig mit angegebener Email
        Return '3': Request nicht vom Typ GET
        Return '4': app.config["TAN_DICT"] leer
-       Return '5': TAN gueltig aber mit anderer Email'''
+       Return '5': TAN gueltig aber mit anderer Email
+       Return '6': TAN gueltig aber falsche Formular'''
     if request.method != "GET":
         print(3)
         return '3'
@@ -155,10 +170,14 @@ def isTanValid() -> str:
         return '4'
     tan = request.args.get('tan')
     email = request.args.get('email')
+    formEnum = request.args.get('formEnum')
     if tan is not None:
         if tan not in app.config[tanDictAppConfigKey]:
             print(0)
             return '0'
+        if app.config[tanDictAppConfigKey][tan][tanDictFormEnumKey] != '-1' and (formEnum is None or formEnum != app.config[tanDictAppConfigKey][tan][tanDictFormEnumKey]):
+            print(6)
+            return '6'
         if app.config[tanDictAppConfigKey][tan][tanDictEmailKey] == "":
             print(1)
             return '1'
@@ -177,35 +196,38 @@ def assignEmailToTan() -> typing.Tuple[typing.Dict[str, typing.Union[str, bool]]
             return jsonify({"success":False, "error": "TAN_DICT muss neu generiert werden"}), 503
         tan = request.form.get('tan')
         email = request.form.get('email')
+        formEnum = request.form.get('formEnum')
         if tan is None:
             return jsonify({"success":False, "error": "Keine TAN Angabe"}), 400
         if email is None:
             return jsonify({"success":False, "error": "Keine e-mail Angabe"}), 400
+        if formEnum is None:
+            return jsonify({"success":False, "error": "Keine formEnum Angabe"}), 400
         email = email.strip()
         if tan in app.config[tanDictAppConfigKey] and app.config[tanDictAppConfigKey][tan][tanDictEmailKey] != "": 
             if app.config[tanDictAppConfigKey][tan][tanDictEmailKey].lower() != email.lower():
                 return jsonify({"success":False, "error": "TAN schon an andere E-mail zugewiesen"}), 409
             return jsonify({"success":True, "message": "TAN schon an diese E-mail zugewiesen"}), 200
         listSizeReached = False
-        with tanListLock:
-            if tan in app.config[tanDictAppConfigKey]:
-                largestCurrentIndex = 0
-                for values in app.config[tanDictAppConfigKey].values():
-                    if values[tanDictIndexKey] > largestCurrentIndex:
-                        largestCurrentIndex += 1
+        if tan in app.config[tanDictAppConfigKey]:
+            largestCurrentIndex = app.config[tanDictAppConfigKey][tanDictLargestIndexKey]
+            with tanListAppLock:
                 app.config[tanDictAppConfigKey][tan][tanDictEmailKey] = email
                 app.config[tanDictAppConfigKey][tan][tanDictIndexKey] = largestCurrentIndex + 1
-                if len([x for x in app.config[tanDictAppConfigKey].values() if x[tanDictEmailKey] != ""]) >= TAN_LIST_SIZE:
-                    listSizeReached = True
-                if not listSizeReached:
-                    if os.path.exists(os.path.join(os.path.dirname(__file__), tanListFileName)):
+                app.config[tanDictAppConfigKey][tan][tanDictFormEnumKey] = formEnum
+                app.config[tanDictAppConfigKey][tanDictLargestIndexKey] = largestCurrentIndex + 1
+            if app.config[tanDictAppConfigKey][tanDictLargestIndexKey] >= TAN_LIST_SIZE:
+                listSizeReached = True
+            if not listSizeReached:
+                if os.path.exists(os.path.join(os.path.dirname(__file__), tanListFileName)):
+                    with tanListFileLock:
                         with open(os.path.join(os.path.dirname(__file__), tanListFileName), "w") as f:
                             json.dump(app.config[tanDictAppConfigKey], f, indent=4)
                             return jsonify({"success":True, "message": "TAN entwertet"}), 200
-                    else:
-                        return jsonify({"success":False, "error": "TAN Liste konnte nicht gefunden werden"}), 404
-            else:
-                return jsonify({"success":False, "error": "TAN nicht vorhanden"}), 404
+                else:
+                    return jsonify({"success":False, "error": "TAN Liste konnte nicht gefunden werden"}), 404
+        else:
+            return jsonify({"success":False, "error": "TAN nicht vorhanden"}), 404
         if listSizeReached:
             _generateTanListAndAssign()
             return jsonify({"success":True, "message": "TAN entwertet"}), 200
@@ -216,23 +238,21 @@ def assignEmailToTan() -> typing.Tuple[typing.Dict[str, typing.Union[str, bool]]
 def _generateTanListAndAssign() -> typing.Tuple[typing.Dict[str, typing.Union[str, bool]], int]:
     TAN_DICT = _generateTanListHelper()
     try:
-        with tanListLock:
-            oldTanDict = {}
-            largestIndex = 0
-            for key, value in app.config[tanDictAppConfigKey].items():
-                if value[tanDictIndexKey] > largestIndex:
-                    largestIndex = value[tanDictIndexKey]
-            for key, value in app.config[tanDictAppConfigKey].items():
-                if value[tanDictIndexKey] == 0 or value[tanDictIndexKey] < largestIndex - TAN_LIST_RENEWAL_RETAIN_SIZE:
-                    continue
-                oldTanDict[key] = value
-            for key in oldTanDict.keys():
-                oldTanDict[key][tanDictIndexKey] -= largestIndex
-            TAN_DICT.update(oldTanDict)
-            out = _generateTanList(TAN_DICT)
-            with open(os.path.join(os.path.dirname(__file__), tanListFileName), "r") as f:
-                app.config[tanDictAppConfigKey] = json.load(f)
-            return out
+        oldTanDict = {}
+        largestIndex = 0
+        if tanDictLargestIndexKey in app.config[tanDictAppConfigKey]:
+            largestIndex = app.config[tanDictAppConfigKey][tanDictLargestIndexKey]
+        for key, value in app.config[tanDictAppConfigKey].items():
+            if key == tanDictLargestIndexKey or value[tanDictIndexKey] <= 0 or value[tanDictIndexKey] < largestIndex - TAN_LIST_RENEWAL_RETAIN_SIZE + 1:
+                continue
+            oldTanDict[key] = value
+        for key in oldTanDict.keys():
+            oldTanDict[key][tanDictIndexKey] -= largestIndex
+        TAN_DICT.update(oldTanDict)
+        out = _generateTanList(TAN_DICT)
+        with tanListAppLock:
+            app.config[tanDictAppConfigKey] = TAN_DICT
+        return out
     except Exception as e:
         return jsonify({"success":False, "error": str(e)}), 500
